+++ packages/terra-graphs-docs/src/example-datasets/dataObjects/Timeline/criticalDataset.js
@@ -20,10 +20,12 @@ const data = [{
     {
       x: new Date(2016, 0, 1, 1, 30).toISOString(),
       content: bobRossIpsum,
+      isCritical: true
     },
     {
       x: new Date(2016, 0, 1, 2, 15).toISOString(),
       content: bobRossIpsum,
+      isCritical: true
     },
     {
       x: new Date(2016, 0, 1, 3, 15).toISOString(),
@@ -32,9 +34,10 @@ const data = [{
     {
       x: new Date(2016, 0, 1, 4, 15).toISOString(),
       content: bobRossIpsum,
+      isCritical: true
     },
     {
-      x: new Date(2016, 0, 1, 5, 15).toISOString(),
+      x: new Date(2016, 0, 1, 8, 15).toISOString(),
       content: bobRossIpsum,
     },
   ],
@@ -51,13 +54,14 @@ const data = [{
     {
       x: new Date(2016, 0, 1, 7, 15).toISOString(),
       content: 'This is custom value of another unit',
+      isCritical: true
     },
     {
       x: new Date(2016, 0, 1, 9, 45).toISOString(),
       content: 'This is custom value of another unit',
     },
     {
-      x: new Date(2016, 0, 1, 12).toISOString(),
+      x: new Date(2016, 0, 1, 15).toISOString(),
       content: 'This is custom value of another unit',
     },
   ],