+++ packages/terra-graphs/src/components/Timeline/Timeline.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import '../Graph.module.scss';
 import './Timeline.module.scss';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 
 const propTypes = {
   /**
@@ -26,8 +27,23 @@ const propTypes = {
 const Timeline = ({
   graphConfig, dataset, graphID, timeout,
 }) => {
+  const [graph, setGraph] = React.useState();
+  const graphLoadedRef = React.useRef();
+  const skipreflowRef = React.useRef();
+
+
+  // creation of canvas
   React.useEffect(() => {
-    const graph = Carbon.api.timeline(graphConfig);
+    if (!graph) {
+      setGraph(Carbon.api.timeline(graphConfig));
+    }
+  }, []);
+
+  // initial dataset load
+  React.useEffect(() => {
+    if (!graph || graphLoadedRef.current) {
+      return;
+    }
     const timeoutIds = [];
 
     if (dataset) {
@@ -48,11 +64,32 @@ const Timeline = ({
         });
       }
     }
-
+    graphLoadedRef.current = true;
     return () => {
       timeoutIds.forEach((id) => { clearTimeout(id); });
     };
-  }, [graphConfig, dataset, timeout]);
+  }, [graph, dataset, timeout]);
+
+  // panning
+  React.useEffect(() => {
+    if (!graph) {
+      return;
+    }
+    if (!skipreflowRef.current) {
+      skipreflowRef.current = true;
+      return;
+    }
+
+    graph.config.axis.x.upperLimit = graphConfig.axis.x.upperLimit;
+    graph.config.axis.x.lowerLimit = graphConfig.axis.x.lowerLimit;
+
+    const newDataset = {
+      panData: utils.deepClone(dataset.panData),
+      eventline: utils.deepClone(dataset.eventline),
+    };
+
+    graph.reflowMultipleDatasets(newDataset);
+  }, [graph, dataset, graphConfig]);
 
   return (
     <div id={`${graphID}-canvasContainer`}>