+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -11,6 +11,6 @@
   * Added data objects and examples for Timeline graph.
   * Added data objects and examples for Pie chart.
   * Added data objects and examples for Combination graph.
-  * Added reflow examples for Line Graph.
   * Added critical and basic example for timeline graph.
+  * Added reflow examples for Line Graph.
   * Added reflow examples for Paired result graph.